Fixed types after refactor
