--- app/product2/_schemas/schemaProductAGeneric.ts
@@ -1,26 +1,0 @@
-// generic schema to be used to build docTemplate schema alongside location shcmeas
-// Try to limit the number of generic schemas whenever possible
-export const schemaGenericA = {
-  checkboxExample: {
-    true: {
-      headerA: "TRUE: Generic schema header location 2",
-      bodyB: ["TRUE: Generic schema body B"],
-      listA: 
-    },
-    false: {
-      bodyA: ["FALSE: Generic schema body A"],
-    },
-  },
-}
-
-export const schemaGenericB = {
-  checkboxExample: {
-    true: {
-      headerA: 
-      bodyA: 
-    },
-    false: {
-      bodyA: ["GB: checkbox is false"],
-    },
-  },
-}
